Size in c

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 	printf("Size of a char: %lu byte(s)\n", sizeof(char));
 	printf("Size of an int: %li byte(s)\n", sizeof(int));
 	printf("Size of a long int: %li byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)(sizeof(d));
+	printf("Size of a long long int: %lu byte(s)\n", (sizeof(d));
 	printf("Size of a float: %ld byte(s)\n", sizeof(float));
 
 	return (0);
